Player Networking
This section will guide you to modify the "Player" prefab. We've first created a player that works as is, but now we are going to modify it so that it works and complies when we use it within PUN environment. The modifications are very light, however the concepts are critical. So this section is very important indeed.
- Transform Synchronization
- Animator Synchronization
- User Input Management
- Camera Control
- Beams Fire Control
- Health Synchronization