Player UI Prefab
- moved the scripts created during the basic Photon tutorial to a separate PhotonTutorial folder

This section will guide you to create the Player UI system. We'll need to show the name of the player and its current health. We'll also need to manage the UI position to follow the players around.

This section has nothing to do with networking per se. However, it raises some very important design patterns, to provide some advanced features revolving around networking and the constraints it introduces in development.

So, the UI is not going to be networked, simply because we don't need to, plenty of other way to go about this and avoid taking up traffic. It's always something to strive for, if you can get away with a feature not to be networked, it's good.

The legitimate question now would be: how can we have an UI for each networked player?

We'll have an UI Prefab with a dedicated PlayerUI script. Our PlayerManager script will hold a reference of this UI Prefab, and will simply instantiate this UI Prefab when the PlayerManager starts, and tells the prefab to follow that very player.
- Creating The UI Prefab
- The PlayerUI Scripts Basics
- Instantiation And Binding With The Player
---- Binding PlayerUI With Player
---- Instantiation
---- Parenting To UI Canvas
- Following The Target Player